cli/pyproject.toml
[project]
name = "knowledge-map-tool"
version = "0.1.0"
description = "A CLI tool for converting markdown files to a Kuzu knowledge graph with simplified schema"
requires-python = ">=3.13"
dependencies = [
    "chonkie[semantic]",
    "click",
    "httpx",
    "hishel",
    "kuzu",
    "python-dotenv",
    "rich",
    "openai",
    "pydantic",
    "jsonschema",
    "watchdog",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "aiohttp>=3.12.15",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
